+++ src/main/java/com/serheev/jwtAuth/repository/RefreshTokenRepository.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.serheev.jwtAuth.model.RefreshToken;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
 
     @Override
-    Optional<RefreshToken> findById(Long id);
+    Optional<RefreshToken> findById(Integer id);
 
     Optional<RefreshToken> findByRefreshToken(String token);
 }
